[refactor] #28 순환참조 해결 위해 getExpectedDuration 서비스 분리
